+++ .github/workflows/testing.yml
@@ -105,6 +105,8 @@ jobs:
           echo "DISPLAY=:99" >> $GITHUB_ENV
 
       - name: Run lint
+        env:
+          PYLINT_EXTRA_ARGS: ${{ matrix.python-version == '3.8' && '--ignored-modules=mcp' || '' }}
         run: npm run lint
 
       - name: Run unit tests
